requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.114.0
python-dotenv==1.0.0